+++ package.json
@@ -1,7 +1,10 @@
 {
   "name": "react-currency-converter-widget",
-  "version": "0.1.0",
-  "private": true,
+  "version": "0.1.8",
+  "description": "This npm package is a UI component for currency conversion in React using TypeScript. It is built with Material-UI, ensuring high compatibility with applications that also use Material-UI.",
+  "private": false,
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "dependencies": {
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
@@ -14,16 +17,21 @@
     "@types/node": "^16.18.112",
     "@types/react": "^18.3.10",
     "@types/react-dom": "^18.3.0",
-    "date-fns": "^4.1.0",
+    "date-fns": "^2.29.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "webpack": "^5.95.0",
+    "webpack-cli": "^5.1.4",
+    "ts-loader": "^9.5.1"
+
+
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build",
+    "build": "tsc",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "predeploy": "npm run build",
@@ -51,5 +59,19 @@
     "@m00nbyte/currency-converter": "^1.0.1",
     "gh-pages": "^6.1.1"
   },
-  "homepage": "https://minnya.github.io/react-currency-converter-widget.git"
+
+  "repository": {
+    "type": "git",
+    "url": "https://minnya.github.io/react-currency-converter-widget.git"
+  },
+  "homepage": "https://minnya.github.io/react-currency-converter-widget.git",
+  "keywords": [
+    "material-ui",
+    "react",
+    "typescript",
+    "currency converter",
+    "npm",
+    "node"
+  ],
+  "license": "MIT"
 }
